# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GameRoomSummaryPhase = typeof GameRoomSummaryPhase[keyof typeof GameRoomSummaryPhase];


export const GameRoomSummaryPhase = {
  lobby: 'lobby',
  playing: 'playing',
  finished: 'finished',
} as const;

export interface GameRoomSummary {
  code: string;
  hostName: string;
  players: number;
  maxPlayers: number;
  phase: GameRoomSummaryPhase;
  createdAt: number;
}

export interface GameRoomList {
  rooms: GameRoomSummary[];
}

export type GameViewPhase = typeof GameViewPhase[keyof typeof GameViewPhase];


export const GameViewPhase = {
  lobby: 'lobby',
  playing: 'playing',
  finished: 'finished',
} as const;

export interface GamePlayer {
  id: string;
  name: string;
  color: string;
  cash: number;
  position: number;
  jailed: boolean;
  bankrupt: boolean;
  properties: number[];
  isHost: boolean;
  jailTurns?: number;
  jailCards?: number;
}

export type GameSpaceType = typeof GameSpaceType[keyof typeof GameSpaceType];


export const GameSpaceType = {
  start: 'start',
  property: 'property',
  transit: 'transit',
  utility: 'utility',
  tax: 'tax',
  chance: 'chance',
  'community-chest': 'community-chest',
  rest: 'rest',
  jail: 'jail',
  'go-to-jail': 'go-to-jail',
} as const;

export interface GameSpace {
  id: number;
  name: string;
  type: GameSpaceType;
  /** @nullable */
  price: number | null;
  /** @nullable */
  rent: number | null;
  /** @nullable */
  group: string | null;
  /** @nullable */
  ownerPlayerId: string | null;
  color: string;
  description: string;
  /**
     * @minimum 0
     * @maximum 5
     */
  buildingLevel: number;
  mortgaged: boolean;
  /** @nullable */
  buildCost: number | null;
  /** @nullable */
  currentRent: number | null;
  developmentRents?: number[];
}

export interface TradeProperty {
  spaceId: number;
  mortgaged: boolean;
}

export type GameTradeStatus = typeof GameTradeStatus[keyof typeof GameTradeStatus];


export const GameTradeStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  cancelled: 'cancelled',
  invalidated: 'invalidated',
} as const;

export interface GameTrade {
  id: string;
  proposerPlayerId: string;
  recipientPlayerId: string;
  offeredCash: number;
  requestedCash: number;
  offeredProperties: TradeProperty[];
  requestedProperties: TradeProperty[];
  status: GameTradeStatus;
  createdTurn: number;
}

export interface GameView {
  code: string;
  phase: GameViewPhase;
  players: GamePlayer[];
  board: GameSpace[];
  trades: GameTrade[];
  /** @nullable */
  currentPlayerId: string | null;
  turnNumber: number;
  lastRoll: number[];
  message: string;
  history: string[];
  /** @nullable */
  winnerPlayerId: string | null;
  /** @nullable */
  myPlayerId: string | null;
  createdAt: number;
  consecutiveDoubles?: number;
  extraRoll?: boolean;
  rollSerial?: number;
}

export interface GameJoinResult {
  game: GameView;
  sessionToken: string;
}

export type PropertyManagementInputAction = typeof PropertyManagementInputAction[keyof typeof PropertyManagementInputAction];


export const PropertyManagementInputAction = {
  build: 'build',
  'sell-building': 'sell-building',
  mortgage: 'mortgage',
  redeem: 'redeem',
} as const;

export interface PropertyManagementInput {
  /** @minLength 16 */
  sessionToken: string;
  /**
     * @minimum 0
     * @maximum 39
     */
  spaceId: number;
  action: PropertyManagementInputAction;
}

export interface TradeProposalInput {
  /** @minLength 16 */
  sessionToken: string;
  /** @minLength 1 */
  recipientPlayerId: string;
  /**
     * @minimum 0
     * @maximum 1000000
     */
  offeredCash: number;
  /**
     * @minimum 0
     * @maximum 1000000
     */
  requestedCash: number;
  /**
     * @maxItems 40
     * @items.minimum 0
     * @items.maximum 39
     */
  offeredPropertyIds: number[];
  /**
     * @maxItems 40
     * @items.minimum 0
     * @items.maximum 39
     */
  requestedPropertyIds: number[];
}

export type TradeResponseInputAction = typeof TradeResponseInputAction[keyof typeof TradeResponseInputAction];


export const TradeResponseInputAction = {
  accept: 'accept',
  reject: 'reject',
  cancel: 'cancel',
} as const;

export interface TradeResponseInput {
  /** @minLength 16 */
  sessionToken: string;
  /** @minLength 1 */
  tradeId: string;
  action: TradeResponseInputAction;
}

export interface GameCreateInput {
  /**
     * @minLength 1
     * @maxLength 18
     */
  playerName: string;
}

export interface GameJoinInput {
  /**
     * @minLength 1
     * @maxLength 18
     */
  playerName: string;
  sessionToken?: string;
}

export interface GameActionInput {
  /** @minLength 16 */
  sessionToken: string;
}

export type JailChoiceInputMethod = typeof JailChoiceInputMethod[keyof typeof JailChoiceInputMethod];


export const JailChoiceInputMethod = {
  pay: 'pay',
  card: 'card',
} as const;

export interface JailChoiceInput {
  /** @minLength 16 */
  sessionToken: string;
  method: JailChoiceInputMethod;
}