# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: games
    description: Shared multiplayer board games
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List public game rooms
      responses:
        "200":
          description: Open public rooms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameRoomList"
    post:
      operationId: createGame
      tags: [games]
      summary: Create a public room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameCreateInput"
      responses:
        "201":
          description: Created room and host session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameJoinResult"
  /games/{code}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get the current room and board state
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, minLength: 4, maxLength: 8 }
      responses:
        "200":
          description: Current game state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameView"
        "404":
          description: Room not found
  /games/{code}/players:
    post:
      operationId: joinGame
      tags: [games]
      summary: Join or reconnect to a public room
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, minLength: 4, maxLength: 8 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameJoinInput"
      responses:
        "200":
          description: Joined room and player session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameJoinResult"
  /games/{code}/start:
    post:
      operationId: startGame
      tags: [games]
      summary: Start a room when the host is ready
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameActionInput"
      responses:
        "200":
          description: Game started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameView"
  /games/{code}/roll:
    post:
      operationId: rollDice
      tags: [games]
      summary: Roll the dice for the active player
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameActionInput"
      responses:
        "200":
          description: Dice rolled and landing resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameView"
  /games/{code}/buy:
    post:
      operationId: buyProperty
      tags: [games]
      summary: Buy the unowned property under the active player
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameActionInput"
      responses:
        "200":
          description: Property purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameView"
  /games/{code}/jail:
    post:
      operationId: leaveJail
      tags: [games]
      summary: Pay the jail fine or use a Get Out of Jail Free card before rolling
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JailChoiceInput"
      responses:
        "200":
          description: Player released from jail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameView"
  /games/{code}/pass:
    post:
      operationId: endTurn
      tags: [games]
      summary: Pass on a property or end the active turn
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameActionInput"
      responses:
        "200":
          description: Turn passed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameView"
  /games/{code}/property:
    post:
      operationId: manageProperty
      tags: [games]
      summary: Build, sell a building, mortgage, or redeem an owned property
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyManagementInput"
      responses:
        "200":
          description: Property managed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameView"
        "400":
          description: Invalid property action
        "403":
          description: Invalid player session
        "404":
          description: Room not found
  /games/{code}/trades:
    post:
      operationId: proposeTrade
      tags: [games]
      summary: Propose a consensual cash and property exchange
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeProposalInput"
      responses:
        "200":
          description: Offer proposed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameView"
        "400":
          description: Invalid trade
        "403":
          description: Invalid player session
        "404":
          description: Room not found
  /games/{code}/trades/respond:
    post:
      operationId: respondTrade
      tags: [games]
      summary: Accept, reject, or cancel an existing offer
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeResponseInput"
      responses:
        "200":
          description: Offer resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameView"
        "400":
          description: Invalid or stale trade
        "403":
          description: Unauthorized response
        "404":
          description: Room or offer not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    GameRoomList:
      type: object
      required: [rooms]
      properties:
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/GameRoomSummary"
    GameRoomSummary:
      type: object
      required: [code, hostName, players, maxPlayers, phase, createdAt]
      properties:
        code: { type: string }
        hostName: { type: string }
        players: { type: integer }
        maxPlayers: { type: integer }
        phase: { type: string, enum: [lobby, playing, finished] }
        createdAt: { type: integer }
    GameJoinResult:
      type: object
      required: [game, sessionToken]
      properties:
        game:
          $ref: "#/components/schemas/GameView"
        sessionToken: { type: string }
    GameView:
      type: object
      required: [code, phase, players, board, trades, currentPlayerId, turnNumber, lastRoll, message, history, winnerPlayerId, myPlayerId, createdAt]
      properties:
        code: { type: string }
        phase: { type: string, enum: [lobby, playing, finished] }
        players:
          type: array
          items:
            $ref: "#/components/schemas/GamePlayer"
        board:
          type: array
          items:
            $ref: "#/components/schemas/GameSpace"
        trades:
          type: array
          items:
            $ref: "#/components/schemas/GameTrade"
        currentPlayerId: { type: ["string", "null"] }
        turnNumber: { type: integer }
        lastRoll:
          type: array
          items: { type: integer }
        message: { type: string }
        history:
          type: array
          items: { type: string }
        winnerPlayerId: { type: ["string", "null"] }
        myPlayerId: { type: ["string", "null"] }
        createdAt: { type: integer }
        consecutiveDoubles: { type: integer }
        extraRoll: { type: boolean }
        rollSerial: { type: integer }
    GamePlayer:
      type: object
      required: [id, name, color, cash, position, jailed, bankrupt, properties, isHost]
      properties:
        id: { type: string }
        name: { type: string }
        color: { type: string }
        cash: { type: integer }
        position: { type: integer }
        jailed: { type: boolean }
        bankrupt: { type: boolean }
        properties:
          type: array
          items: { type: integer }
        isHost: { type: boolean }
        jailTurns: { type: integer }
        jailCards: { type: integer }
    GameSpace:
      type: object
      required: [id, name, type, price, rent, group, ownerPlayerId, color, description, buildingLevel, mortgaged, buildCost, currentRent]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string, enum: [start, property, transit, utility, tax, chance, community-chest, rest, jail, go-to-jail] }
        price: { type: ["integer", "null"] }
        rent: { type: ["integer", "null"] }
        group: { type: ["string", "null"] }
        ownerPlayerId: { type: ["string", "null"] }
        color: { type: string }
        description: { type: string }
        buildingLevel: { type: integer, minimum: 0, maximum: 5 }
        mortgaged: { type: boolean }
        buildCost: { type: ["integer", "null"] }
        currentRent: { type: ["integer", "null"] }
        developmentRents:
          type: array
          items: { type: integer }
    GameTrade:
      type: object
      required: [id, proposerPlayerId, recipientPlayerId, offeredCash, requestedCash, offeredProperties, requestedProperties, status, createdTurn]
      properties:
        id: { type: string }
        proposerPlayerId: { type: string }
        recipientPlayerId: { type: string }
        offeredCash: { type: integer }
        requestedCash: { type: integer }
        offeredProperties:
          type: array
          items:
            $ref: "#/components/schemas/TradeProperty"
        requestedProperties:
          type: array
          items:
            $ref: "#/components/schemas/TradeProperty"
        status: { type: string, enum: [pending, accepted, rejected, cancelled, invalidated] }
        createdTurn: { type: integer }
    TradeProperty:
      type: object
      required: [spaceId, mortgaged]
      properties:
        spaceId: { type: integer }
        mortgaged: { type: boolean }
    PropertyManagementInput:
      type: object
      required: [sessionToken, spaceId, action]
      properties:
        sessionToken: { type: string, minLength: 16 }
        spaceId: { type: integer, minimum: 0, maximum: 39 }
        action: { type: string, enum: [build, sell-building, mortgage, redeem] }
    TradeProposalInput:
      type: object
      required: [sessionToken, recipientPlayerId, offeredCash, requestedCash, offeredPropertyIds, requestedPropertyIds]
      properties:
        sessionToken: { type: string, minLength: 16 }
        recipientPlayerId: { type: string, minLength: 1 }
        offeredCash: { type: integer, minimum: 0, maximum: 1000000 }
        requestedCash: { type: integer, minimum: 0, maximum: 1000000 }
        offeredPropertyIds:
          type: array
          maxItems: 40
          items: { type: integer, minimum: 0, maximum: 39 }
        requestedPropertyIds:
          type: array
          maxItems: 40
          items: { type: integer, minimum: 0, maximum: 39 }
    TradeResponseInput:
      type: object
      required: [sessionToken, tradeId, action]
      properties:
        sessionToken: { type: string, minLength: 16 }
        tradeId: { type: string, minLength: 1 }
        action: { type: string, enum: [accept, reject, cancel] }
    GameCreateInput:
      type: object
      required: [playerName]
      properties:
        playerName: { type: string, minLength: 1, maxLength: 18 }
    GameJoinInput:
      type: object
      required: [playerName]
      properties:
        playerName: { type: string, minLength: 1, maxLength: 18 }
        sessionToken: { type: string }
    GameActionInput:
      type: object
      required: [sessionToken]
      properties:
        sessionToken: { type: string, minLength: 16 }
    JailChoiceInput:
      type: object
      required: [sessionToken, method]
      properties:
        sessionToken: { type: string, minLength: 16 }
        method: { type: string, enum: [pay, card] }