# lib/api-zod/src/generated/types/gameSpace.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameSpaceType } from './gameSpaceType';

export interface GameSpace {
  id: number;
  name: string;
  type: GameSpaceType;
  /** @nullable */
  price: number | null;
  /** @nullable */
  rent: number | null;
  /** @nullable */
  group: string | null;
  /** @nullable */
  ownerPlayerId: string | null;
  color: string;
  description: string;
  /**
     * @minimum 0
     * @maximum 5
     */
  buildingLevel: number;
  mortgaged: boolean;
  /** @nullable */
  buildCost: number | null;
  /** @nullable */
  currentRent: number | null;
  developmentRents?: number[];
}