# lib/api-zod/src/generated/types/gameTrade.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameTradeStatus } from './gameTradeStatus';
import type { TradeProperty } from './tradeProperty';

export interface GameTrade {
  id: string;
  proposerPlayerId: string;
  recipientPlayerId: string;
  offeredCash: number;
  requestedCash: number;
  offeredProperties: TradeProperty[];
  requestedProperties: TradeProperty[];
  status: GameTradeStatus;
  createdTurn: number;
}