# lib/api-zod/src/generated/types/gameView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GamePlayer } from './gamePlayer';
import type { GameSpace } from './gameSpace';
import type { GameTrade } from './gameTrade';
import type { GameViewPhase } from './gameViewPhase';

export interface GameView {
  code: string;
  phase: GameViewPhase;
  players: GamePlayer[];
  board: GameSpace[];
  trades: GameTrade[];
  /** @nullable */
  currentPlayerId: string | null;
  turnNumber: number;
  lastRoll: number[];
  message: string;
  history: string[];
  /** @nullable */
  winnerPlayerId: string | null;
  /** @nullable */
  myPlayerId: string | null;
  createdAt: number;
  consecutiveDoubles?: number;
  extraRoll?: boolean;
  rollSerial?: number;
}