# lib/api-zod/src/generated/types/propertyManagementInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyManagementInputAction } from './propertyManagementInputAction';

export interface PropertyManagementInput {
  /** @minLength 16 */
  sessionToken: string;
  /**
     * @minimum 0
     * @maximum 39
     */
  spaceId: number;
  action: PropertyManagementInputAction;
}