# lib/api-zod/src/generated/types/tradeProposalInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TradeProposalInput {
  /** @minLength 16 */
  sessionToken: string;
  /** @minLength 1 */
  recipientPlayerId: string;
  /**
     * @minimum 0
     * @maximum 1000000
     */
  offeredCash: number;
  /**
     * @minimum 0
     * @maximum 1000000
     */
  requestedCash: number;
  /**
     * @maxItems 40
     * @items.minimum 0
     * @items.maximum 39
     */
  offeredPropertyIds: number[];
  /**
     * @maxItems 40
     * @items.minimum 0
     * @items.maximum 39
     */
  requestedPropertyIds: number[];
}