# lib/api-zod/src/generated/types/tradeResponseInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeResponseInputAction } from './tradeResponseInputAction';

export interface TradeResponseInput {
  /** @minLength 16 */
  sessionToken: string;
  /** @minLength 1 */
  tradeId: string;
  action: TradeResponseInputAction;
}